+++ inst/include/dust/gpu/filter.hpp
@@ -11,6 +11,7 @@ namespace filter {
 template <typename T>
 std::vector<typename T::real_type>
 filter(T * obj,
+       size_t step_end,
        filter_state_device<typename T::real_type>& state,
        bool save_trajectories,
        std::vector<size_t> step_snapshot) {
@@ -28,7 +29,12 @@ filter(T * obj,
 
   if (save_trajectories) {
     state.trajectories.resize(obj->n_state(), n_particles, n_data);
-    state.trajectories.store_values(obj->device_state_selected());
+
+    const auto step_first_data = obj->data().begin()->first;
+    if (obj->step() <= step_first_data) {
+      state.trajectories.store_values(obj->device_state_selected());
+    }
+
     state.trajectories.advance();
   }
 
@@ -38,17 +44,21 @@ filter(T * obj,
     state.snapshots.resize(obj->n_state_full(), n_particles, step_snapshot);
   }
 
-  for (auto & d : obj->data()) {
-    // MODEL UPDATE
-    obj->run(d.first);
+  auto d = obj->data().cbegin();
+  const auto d_end = obj->data().cend();
 
-    // SAVE HISTORY (async)
-    if (save_trajectories) {
-      state.trajectories.store_values(obj->device_state_selected());
-    }
+  while (d->first <= obj->step() && d != d_end) {
+    d++;
+    state.trajectories.advance();
+  }
+
+  for (; d != d_end && obj->step() < step_end; ++d) {
+    const auto step = d->first;
+    // MODEL UPDATE
+    obj->run(step);
 
     // COMPARISON FUNCTION
-    obj->compare_data(weights.weights(), d.second);
+    obj->compare_data(weights.weights(), d->second);
 
     // SCALE WEIGHTS
     weights.scale_log_weights(log_likelihood);
@@ -57,14 +67,15 @@ filter(T * obj,
     // Normalise the weights and calculate cumulative sum for resample
     obj->resample(weights.weights(), scan);
 
-    // SAVE HISTORY ORDER
+    // SAVE HISTORY (async)
     if (save_trajectories) {
+      state.trajectories.store_values(obj->device_state_selected());
       state.trajectories.store_order(obj->filter_kappa());
       state.trajectories.advance();
     }
 
     // SAVE SNAPSHOT
-    if (save_snapshots && state.snapshots.is_snapshot_step(d.first)) {
+    if (save_snapshots && state.snapshots.is_snapshot_step(step)) {
       state.snapshots.store(obj->device_state_full());
       state.snapshots.advance();
     }