+++ inst/include/dust/r/dust.hpp
@@ -446,29 +446,33 @@ cpp11::sexp save_snapshots(const filter_state& snapshots, const T *obj,
   return(r_snapshots);
 }
 
-// TODO: this should be rewritten to return a list of 3 cpp1::sexp not
-// passing two as references.
 template <typename T>
-cpp11::writable::doubles run_filter(T * obj,
-                                    cpp11::sexp& r_trajectories,
-                                    cpp11::sexp& r_snapshots,
-                                    std::vector<size_t>& step_snapshot,
-                                    bool save_trajectories) {
+cpp11::sexp run_filter(T * obj, size_t step,
+                       std::vector<size_t>& step_snapshot,
+                       bool save_trajectories) {
   typename T::filter_state_type filter_state;
   cpp11::writable::doubles log_likelihood =
-    dust::filter::filter(obj, filter_state, save_trajectories, step_snapshot);
+    dust::filter::filter(obj, step, filter_state, save_trajectories,
+                         step_snapshot);
+  cpp11::sexp r_trajectories, r_snapshots;
   if (save_trajectories) {
     r_trajectories = dust::r::save_trajectories(filter_state.trajectories, obj);
   }
   if (!step_snapshot.empty()) {
-    r_snapshots = dust::r::save_snapshots(filter_state.snapshots, obj, step_snapshot);
+    r_snapshots = dust::r::save_snapshots(filter_state.snapshots, obj,
+                                          step_snapshot);
   }
-  return log_likelihood;
+
+  using namespace cpp11::literals;
+  return cpp11::writable::list({"log_likelihood"_nm = log_likelihood,
+                                "trajectories"_nm = r_trajectories,
+                                "snapshots"_nm = r_snapshots});
 }
 
 template <typename T, typename std::enable_if<!std::is_same<dust::no_data, typename T::data_type>::value, int>::type = 0>
-cpp11::sexp dust_filter(SEXP ptr, bool save_trajectories,
-                        cpp11::sexp r_step_snapshot) {
+cpp11::sexp dust_filter(SEXP ptr, SEXP r_step_end, bool save_trajectories,
+                        cpp11::sexp r_step_snapshot,
+                        cpp11::sexp min_log_likelihood) {
   T *obj = cpp11::as_cpp<cpp11::external_pointer<T>>(ptr).get();
   obj->check_errors();
 
@@ -476,18 +480,27 @@ cpp11::sexp dust_filter(SEXP ptr, bool save_trajectories,
     cpp11::stop("Data has not been set for this object");
   }
 
+  size_t step_end = std::numeric_limits<size_t>::max();
+  if (r_step_end != R_NilValue) {
+    step_end = cpp11::as_cpp<int>(r_step_end);
+    dust::r::validate_size(step_end, "step_end");
+    if (step_end <= obj->step()) {
+      cpp11::stop("'step_end' must be larger than curent step (%d; given %d)",
+                  obj->step(), step_end);
+    }
+    if (obj->data().find(step_end) == obj->data().end()) {
+      cpp11::stop("'step_end' was not found in data (was given %d)", step_end);
+    }
+  }
+
   std::vector<size_t> step_snapshot =
     dust::r::check_step_snapshot(r_step_snapshot, obj->data());
 
-  cpp11::sexp r_trajectories, r_snapshots;
-  cpp11::writable::doubles log_likelihood =
-    run_filter<T>(obj, r_trajectories, r_snapshots, step_snapshot,
-                  save_trajectories);
+  if (min_log_likelihood != R_NilValue) {
+    cpp11::stop("min_log_likelihood not yet supported");
+  }
 
-  using namespace cpp11::literals;
-  return cpp11::writable::list({"log_likelihood"_nm = log_likelihood,
-                                "trajectories"_nm = r_trajectories,
-                                "snapshots"_nm = r_snapshots});
+  return run_filter<T>(obj, step_end, step_snapshot, save_trajectories);
 }
 
 // Based on the value of the data_type in the underlying model class we
@@ -510,8 +523,9 @@ cpp11::sexp dust_compare_data(SEXP ptr) {
 }
 
 template <typename T, typename std::enable_if<std::is_same<dust::no_data, typename T::data_type>::value, int>::type = 0>
-cpp11::sexp dust_filter(SEXP ptr, bool save_trajectories,
-                        cpp11::sexp step_snapshot) {
+cpp11::sexp dust_filter(SEXP ptr, SEXP step_end, bool save_trajectories,
+                        cpp11::sexp step_snapshot,
+                        cpp11::sexp min_log_likelihood) {
   disable_method("filter");
   return R_NilValue; // #nocov never gets here
 }