+++ inst/template/dust_methods.cpp
@@ -63,9 +63,14 @@ SEXP dust_{{target}}_{{name}}_compare_data(SEXP ptr) {
   return dust::r::dust_compare_data<model_{{target}}>(ptr);
 }
 
-SEXP dust_{{target}}_{{name}}_filter(SEXP ptr, bool save_trajectories,
-                          cpp11::sexp step_snapshot) {
-  return dust::r::dust_filter<model_{{target}}>(ptr, save_trajectories, step_snapshot);
+SEXP dust_{{target}}_{{name}}_filter(SEXP ptr, SEXP step_end,
+                                     bool save_trajectories,
+                                     cpp11::sexp step_snapshot,
+                                     cpp11::sexp min_log_likelihood) {
+  return dust::r::dust_filter<model_{{target}}>(ptr, step_end,
+                                                save_trajectories,
+                                                step_snapshot,
+                                                min_log_likelihood);
 }
 
 void dust_{{target}}_{{name}}_set_n_threads(SEXP ptr, int n_threads) {